Add DPI scaling variable to cfgmgr Fill in DPI scaling combo box
git-svn-id: https://www.williamfeely.info/svn/dxgl@409 8a90861a-4eca-46d5-b744-240ff16d0c4d

+++ cfgmgr/cfgmgr.cpp
@@ -106,7 +106,7 @@ DWORD ReadDWORD(HKEY hKey, DWORD original, DWORD &mask, LPCTSTR value)
 
 float ReadFloat(HKEY hKey, float original, float &mask, LPCTSTR value)
 {
-	DWORD dwOut;
+	float dwOut;
 	DWORD sizeout = 4;
 	DWORD regdword = REG_DWORD;
 	LSTATUS error = RegQueryValueEx(hKey,value,NULL,&regdword,(LPBYTE)&dwOut,&sizeout);
